lopper/assists: allow chained command line assists
When assists are autorun from the command line, the name of the
module/assists follows the first -- on the command line.

In some situations we may want to chain assists from the
command line. So we make the processing more flexible and allow
multiple command line modules to be specified if they are
separated by --

It is the module/assists responsibility to pass information
through the pipeline.

Signed-off-by: Bruce Ashfield <bruce.ashfield@xilinx.com>