Optional mode
None : Only required
Define : Adding a #define with the name of the property for each present 
optional properties
Boolean : Write down all optional properties alongside a boolean named 
by the property name + '_p' (for présence). This boolean is set to false 
if property not present and the property value set to it's default value 
(NULL for pointer and 0 for other)

With stm32mp15 dt, there is a pointer that is not correctly written down 
in boolean mode.